+++ backend/api/src/main/java/com/mycompany/pinterior/dao/PinDao.java
@@ -16,13 +16,13 @@ public interface PinDao {
 	public int insert(Pin pin);
 	
 	// 핀 단건 조회 
-	Pin selectById(int pinId);
+	Pin selectById(Long pinId);
 
 	// 핀 수정 
 	public int update(Pin pin);
 	
 	// 핀 삭제
-	public int deleteByPinId(int pinId);
+	public int deleteByPinId(Long pinId);
 	
 	// ===== 김효: 전체 핀 목록 조회 =====
 	List<PinSummaryDto> selectList(@Param("cursorId") Long cursorId, @Param("size") int size);
