+++ backend/api/src/main/java/com/mycompany/pinterior/dto/PinUpdateResponseDto.java
@@ -7,12 +7,12 @@ import lombok.Data;
 
 @Data
 public class PinUpdateResponseDto {
-	private int pinId;
-	private int userId;
+	private Long pinId;
+	private Long userId;
 	private String title;
 	private String description;
 	private String linkUrl;
 	private List<String> tags;
-	private int boardId;
+	private Long boardId;
 	private Date updatedAt;
 }
